backend/requirements.txt
duckdb==1.1.3
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
pyarrow==18.1.0
python-dotenv==1.0.1
